Added support for linked meshes
- linking LOD meshes instead of objects make it possible to move objects and perform anastyloses
